requirements.txt
bs4
yaspin
pysocks
requests
streamlit
python-dotenv
langchain-openai
langchain-anthropic
langchain_community
langchain_google_genai